Versão Beta 0.1
Essa versão pode apresentar erros

+++ Escritorio/PG2_Escritorio1.php
@@ -29,19 +29,33 @@ $estoqueMasculina = isset($_POST['estoqueMasculina']) ? $_POST['estoqueMasculina
 $numeroPedidoSplit = str_split($numeroPedido,1);
 
 $idPedidos = $numeroPedido ."-". $dataEntrega;
+$pdf = null;
 
 //imagem 
+
 if(isset($_FILES['imagem']) && !empty($_FILES['imagem'])){
-    $imagem = "../imagem/". $idPedidos . '.png' ;
-    move_uploaded_file($_FILES['imagem']['tmp_name'], "../imagem/". $idPedidos . '.png');
+
+    if(!file_exists("../imagem/". $idPedidos . '.png')){
+
+        $imagem = "../imagem/". $idPedidos . '.png';
+        move_uploaded_file($_FILES['imagem']['tmp_name'], "../imagem/". $idPedidos . '.png');
+    }
+    $imagem = "../imagem/". $idPedidos . '.png';
 }
+
 
 // Quardando PDF 
-if(isset($_FILES['pdf']) && !empty($_FILES['pdf'])){
-    $pdf = "./pedidos/PDF/" . $idPedidos . '.pdf';
-    move_uploaded_file($_FILES['pdf']['tmp_name'], "./pedidos/PDF/" . $idPedidos . '.pdf' );
-}
+if (isset($_FILES['pdf']) && $_FILES['pdf']['error'] === 0) {
 
+    if(!file_exists("./pedidos/PDF/" . $idPedidos . '.pdf')){
+
+        $pdf = "./pedidos/PDF/" . $idPedidos . '.pdf';
+        move_uploaded_file($_FILES['pdf']['tmp_name'], $pdf);
+    }
+    $pdf = "./pedidos/PDF/" . $idPedidos . '.pdf';
+} else {
+    $pdf = "./pedidos/PDF/semEtiqueta.pdf";
+}
 
 // Outros Clientes 
 if($cliente == 'showroom'){
@@ -67,32 +81,50 @@ if($cliente == 'Outros'){
     VALUES ('$idPedidos','$cliente', '$nomePedido', '$f', '$m', '$descricao_Pedido', '$descricaoAlianca','$largura', '$gravacao_inter', '$gravacao_exter', '$outrosClientes','$imagem' , '$estoqueFeminina' ,'$estoqueMasculina', '$comPedra' , '$semPedra' , 'naoTemPdf' , '$dataEntrega')");
     
 }
+
 if($cliente == 'Mercado_Livre' && $numeroPedidoSplit[1] == "F"){
     
     $idPf = str_replace("PF","",$_POST['numeroPedido']); //Contador para o banco
-    // passando pro banco de dados
-    $dadosp = mysqli_query($conectar, "INSERT INTO pedidosp 
-    (contadorpf,idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca,largura, gravacaoInterna, gravacaoExterna ,imagem,parEstoqueF,parEstoqueM, PedraF,PedraM,pdf,data_digitada) 
-    VALUES ($idPf,'$idPedidos','$cliente', '$nomePedido', '$f', '$m', '$descricao_Pedido', '$descricaoAlianca','$largura', '$gravacao_inter', '$gravacao_exter' ,'$imagem' , '$estoqueFeminina' ,'$estoqueMasculina', '$comPedra' , '$semPedra' , '$pdf' , '$dataEntrega' )");
     
+    // passando pro banco de dados
+    $dados = $conectar->prepare("INSERT INTO pedidosp 
+    (contadorpf, idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca, largura, gravacaoInterna, gravacaoExterna, imagem, parEstoqueF, parEstoqueM, PedraF, PedraM, pdf, data_digitada) 
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+
+    $dados->bind_param("isssiissssssssssss", $idPf, $idPedidos, $cliente, $nomePedido, $f, $m, $descricao_Pedido, 
+                            $descricaoAlianca,$largura, $gravacao_inter, $gravacao_exter, $imagem , 
+                            $estoqueFeminina ,$estoqueMasculina, $comPedra , $semPedra , $pdf , $dataEntrega );
 }
+
 if($cliente == 'Mercado_Livre' && $numeroPedidoSplit[1] == 'G'){
 
-    $idPg = str_replace("PG","",$_POST['numeroPedido']); 
+    $idPg = str_replace("PG","",$_POST['numeroPedido']);
+
     // passando pro banco de dados
-    $dadosg = mysqli_query($conectar, "INSERT INTO pedidospg
-    (contadorpg,idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca,largura, gravacaoInterna, gravacaoExterna,imagem,parEstoqueF,parEstoqueM,PedraF,PedraM, pdf,data_digitada) 
-    VALUES ('$idPg','$idPedidos','$cliente', '$nomePedido', '$f', '$m', '$descricao_Pedido', '$descricaoAlianca','$largura', '$gravacao_inter', '$gravacao_exter','$imagem' , '$estoqueFeminina' ,'$estoqueMasculina', '$comPedra' , '$semPedra' , '$pdf' , '$dataEntrega' )");
+     $dados = $conectar->prepare("INSERT INTO pedidospg 
+    (contadorpg, idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca, largura, gravacaoInterna, gravacaoExterna, imagem, parEstoqueF, parEstoqueM, PedraF, PedraM, pdf, data_digitada) 
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+
+    $dados->bind_param("isssiissssssssssss", $idPg, $idPedidos, $cliente, $nomePedido, $f, $m, $descricao_Pedido, 
+                            $descricaoAlianca,$largura, $gravacao_inter, $gravacao_exter, $imagem , 
+                            $estoqueFeminina ,$estoqueMasculina, $comPedra , $semPedra , $pdf , $dataEntrega );
      
 }
+
 if($cliente == 'Mercado_Livre' && $numeroPedidoSplit[1] == 'E'){
 
     $idPe = str_replace("PE","",$_POST['numeroPedido']);
+
     // passando pro banco de dados
-    $dadose = mysqli_query($conectar, "INSERT INTO pedidospe
-    (contadorpe,idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca,largura, gravacaoInterna, gravacaoExterna,imagem,parEstoqueF,parEstoqueM,PedraF,PedraM,pdf,data_digitada) 
-    VALUES ('$idPe','$idPedidos','$cliente', '$nomePedido', '$f', '$m', '$descricao_Pedido', '$descricaoAlianca','$largura', '$gravacao_inter', '$gravacao_exter', '$imagem' , '$estoqueFeminina' ,'$estoqueMasculina', '$comPedra' , '$semPedra' , '$pdf' , '$dataEntrega' )");
-    
+    $dados = $conectar->prepare("INSERT INTO pedidospe 
+    (contadorpe, idpedidos, cliente, nomePedido, numF, numeM, descricaoPedido, descricaoAlianca, largura, gravacaoInterna, gravacaoExterna, imagem, parEstoqueF, parEstoqueM, PedraF, PedraM, pdf, data_digitada) 
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)");
+
+    $dados->bind_param("isssiissssssssssss", $idPe, $idPedidos, $cliente, $nomePedido, $f, $m, $descricao_Pedido, 
+                            $descricaoAlianca,$largura, $gravacao_inter, $gravacao_exter, $imagem , 
+                            $estoqueFeminina ,$estoqueMasculina, $comPedra , $semPedra , $pdf , $dataEntrega );
 }
 
+$dados->execute();
+$dados->close();
 ?>
